+++ ikarus/utils/observer/nonLinearSolverLogger.hh
@@ -21,10 +21,10 @@ public:
         rNorm = 0.0;
         sNorm = 0.0;
         spdlog::info("Non-linear solver started:");
-        spdlog::info("  i ResidualNorm CorrectionNorm ScalarSubsidiaryNorm");
+        spdlog::info("  i ResidualNorm CorrectionNorm ScalarSubsidiaryNorm lambda");
         break;
       case NonLinearSolverMessages::ITERATION_ENDED:
-        spdlog::info("{:>3d} {:>12.2e} {:14.2e} {:20.2e}", iters, rNorm, dNorm, sNorm);
+        spdlog::info("{:>3d} {:>12.2e} {:14.2e} {:20.2e} {:8.2e}", iters, rNorm, dNorm, sNorm, lambda);
         ++iters;
         break;
       default:
@@ -39,6 +39,9 @@ public:
         break;
       case NonLinearSolverMessages::CORRECTIONNORM_UPDATED:
         dNorm = val;
+        break;
+      case NonLinearSolverMessages::SOLUTION_CHANGED:
+        lambda = val;
         break;
       case NonLinearSolverMessages::SCALARSUBSIDIARY_UPDATED:
         sNorm = val;
@@ -68,4 +71,5 @@ private:
   double dNorm{0};
   double rNorm{0};
   double sNorm{0};
+  double lambda{0};
 };